refactor: make the file monitor a package
Its two on-disk formats were flat modules named for what they hold. Make
the monitor a package shaped like what it writes, the way the wandb
monitor is already a package:

    monitors/file/monitor.py            FileMonitor
    monitors/file/traces/__init__.py    where the stream and its indexes sit
    monitors/file/traces/index.py       the episode index
    monitors/file/traces/update.py      the trace updates

`monitors.file.FileMonitor` still imports from the package root.
