+++ src/prime_rl/monitors/file/__init__.py
@@ -1,0 +1,3 @@
+from prime_rl.monitors.file.monitor import FileMonitor
+
+__all__ = ["FileMonitor"]
